Worked over the code for memory and functionality.
Put all static strings in Flash memory, and conditionally compiled
out the Serial print lines. I removed the extensions from the
sensors and recalibrated the threshold for edge detection.

+++ src/Move.h
@@ -1,7 +1,7 @@
 // myRobotEdge Arduino sketch
 //
 // Copyright (c) 2012 Michael Margolis
-// Copyright (c) 2013 Dave Sieh
+// Copyright (c) 2013,2014 Dave Sieh
 //
 // See LICENSE.txt for details.
 
@@ -35,7 +35,7 @@ class Move {
   void faster();
   int getState();
   void rotate(int angle);
-  void timedMove(int direction, int duration);
+  void timedMove(int direction, long duration);
 
  private:
   int changeMoveState(int newState);
